Revert "lib: print to stdout/stderr directly instead of using console"
This reverts commit 2b24ffae2240163a74ae11e49ee198e98abb07dc.

Fixes: https://github.com/nodejs/node/issues/27819
