+++ lib/fs.js
@@ -75,8 +75,7 @@ const {
   validateOffsetLengthRead,
   validateOffsetLengthWrite,
   validatePath,
-  warnOnNonPortableTemplate,
-  handleErrorFromBinding
+  warnOnNonPortableTemplate
 } = require('internal/fs/utils');
 const {
   CHAR_FORWARD_SLASH,
@@ -116,6 +115,23 @@ function showTruncateDeprecation() {
       'fs.ftruncate with a file descriptor instead.',
       'DeprecationWarning', 'DEP0081');
     truncateWarn = false;
+  }
+}
+
+function handleErrorFromBinding(ctx) {
+  if (ctx.errno !== undefined) {  // libuv error numbers
+    const err = uvException(ctx);
+    // eslint-disable-next-line no-restricted-syntax
+    Error.captureStackTrace(err, handleErrorFromBinding);
+    throw err;
+  }
+  if (ctx.error !== undefined) {  // Errors created in C++ land.
+    // TODO(joyeecheung): currently, ctx.error are encoding errors
+    // usually caused by memory problems. We need to figure out proper error
+    // code(s) for this.
+    // eslint-disable-next-line no-restricted-syntax
+    Error.captureStackTrace(ctx.error, handleErrorFromBinding);
+    throw ctx.error;
   }
 }
